Profile Picture logic added and tested in backend
