+++ config/local.js
@@ -1,34 +1,7 @@
 const passport = require('passport');
 const LocalStrategy = require('passport-local');
 const User = require('../schema/user');
-const multer=require('multer')
 const path=require('path')
-//MULTER 
-const storage=multer.diskStorage({
-  destination:'./public/uploads',
-  filename:function(req,file,cb){
-      cb(null,file.fieldname+"-"+Date.now()+path.extname(file.originalname))
-  }
-})
-const upload=multer({
-  storage:storage,
-  limits:{fileSize:1000000},
-  fileFilter:function(req,res,cb){
-    checkFileType(file,cb);
-  }
-}).single('myImage')
-
-function checkFileType(file,cb){
-  const fileTypes=/jpeg|jpg|png|gif/;
-  const extname=fileTypes.test(path.extname(file.originalname).toLowerCase())
-  const mimetype=fileTypes.test(file.mimetype);
-
-  if(mimetype && extname){
-    return cb(null,true);
-  }else{
-    cb('Error:Images only')
-  }
-}
 
 passport.serializeUser(function(user, done) {
   done(null, user.id);
